financials/pyproject.toml
[project]
name = "financials"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "pandas>=2.2.3",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
]
